requirements.txt
selenium==4.11.2
pandas==2.0.3
python-dotenv==1.0.0
supabase==1.0.3
webdriver-manager==4.0.0
